# remove inline declarations and refactoring