--- a/event-emitter-benchmark.asd
+++ b/event-emitter-benchmark.asd
@@ -1,0 +1,6 @@
+(defsystem "event-emitter-benchmark"
+  :depends-on ("event-emitter")
+  :serial t
+  :components ((:file "t/benchmark"))
+  :perform (test-op :after (o c)
+                    (uiop:symbol-call :event-emitter-benchmark :main)))