I removed the inline declaration and refactored the code.
I couldn't think of any particular advantages to this inline declaration, but as a disadvantage of inline declarations, debugging information is lost and the code becomes less readable.
I also think it is better to optimize the internal implementation rather than inlining the exported functions.