@cxxxr 2024-06-07
A benchmark has been added.
Comparing before and after refactoring, there was no effect from the inline declaration.

before
```
Evaluation took:
  2.750 seconds of real time
  2.741348 seconds of total run time (2.681467 user, 0.059881 system)
  [ Real times consist of 0.154 seconds GC time, and 2.596 seconds non-GC time. ]
  [ Run times consist of 0.152 seconds GC time, and 2.590 seconds non-GC time. ]
  99.67% CPU
  9 lambdas converted
  6,019,819,039 processor cycles
  5,917,078,928 bytes consed
```

after
```
Evaluation took:
  2.775 seconds of real time
  2.764296 seconds of total run time (2.702289 user, 0.062007 system)
  [ Real times consist of 0.150 seconds GC time, and 2.625 seconds non-GC time. ]
  [ Run times consist of 0.153 seconds GC time, and 2.612 seconds non-GC time. ]
  99.60% CPU
  9 lambdas converted
  6,074,091,252 processor cycles
  5,917,121,456 bytes consed
```

```