@cxxxr 2024-06-07
Looking at examples of its use in jsonrpc, it is referenced from multiple threads, so it is necessary to make it thread-safe.
This is a preliminary refactoring for that purpose.