@fukamachi 2024-06-07
Fine, as I don't have a chance to use this library anymore.
I suppose it was not for anything about performance since event-emitter is quite slow and does not care about speed.